requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17



# LangGraph & LLM
langgraph==0.2.60
langchain-core==0.3.29
langchain-google-genai==2.0.8
google-genai>=1.0.0,<2

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary>=2.9.0

# LinkedIn & HTTP
httpx==0.28.1
authlib==1.3.0

# Scheduling
apscheduler==3.10.4

# Config & utils
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Image (Gemini image generation save as PNG)
Pillow>=10.0.0

# Logging & dev
structlog==24.4.0
